Fix up enterprise compatibility for gateways
