+++ agent/proxycfg/state_test.go
@@ -920,222 +920,6 @@ func TestState_WatchesAndUpdates(t *testing.T) {
 				},
 			},
 		},
-		"terminating-gateway-handle-update": testCase{
-			ns: structs.NodeService{
-				Kind:    structs.ServiceKindTerminatingGateway,
-				ID:      "terminating-gateway",
-				Service: "terminating-gateway",
-				Address: "10.0.1.1",
-			},
-			sourceDC: "dc1",
-			stages: []verificationStage{
-				verificationStage{
-					requiredWatches: map[string]verifyWatchRequest{
-						rootsWatchID: genVerifyRootsWatch("dc1"),
-						gatewayServicesWatchID: genVerifyServiceSpecificRequest(gatewayServicesWatchID,
-							"terminating-gateway", "", "dc1", false),
-					},
-					events: []cache.UpdateEvent{
-						rootWatchEvent(),
-						cache.UpdateEvent{
-							CorrelationID: gatewayServicesWatchID,
-							Result: &structs.IndexedGatewayServices{
-								Services: structs.GatewayServices{
-									{
-										Service: structs.NewServiceID("db", nil),
-										Gateway: structs.NewServiceID("terminating-gateway", nil),
-									},
-								},
-							},
-							Err: nil,
-						},
-					},
-					verifySnapshot: func(t testing.TB, snap *ConfigSnapshot) {
-						require.True(t, snap.Valid(), "gateway with service list is valid")
-						require.Len(t, snap.TerminatingGateway.WatchedServices, 1)
-					},
-				},
-				verificationStage{
-					events: []cache.UpdateEvent{
-						cache.UpdateEvent{
-							CorrelationID: gatewayServicesWatchID,
-							Result: &structs.IndexedGatewayServices{
-								Services: structs.GatewayServices{
-									{
-										Service: structs.NewServiceID("db", nil),
-										Gateway: structs.NewServiceID("terminating-gateway", nil),
-									},
-									{
-										Service: structs.NewServiceID("billing", nil),
-										Gateway: structs.NewServiceID("terminating-gateway", nil),
-									},
-								},
-							},
-							Err: nil,
-						},
-					},
-					verifySnapshot: func(t testing.TB, snap *ConfigSnapshot) {
-						db := structs.NewServiceID("db", nil)
-						billing := structs.NewServiceID("billing", nil)
-
-						require.True(t, snap.Valid(), "gateway with service list is valid")
-						require.Len(t, snap.TerminatingGateway.WatchedServices, 2)
-						require.Contains(t, snap.TerminatingGateway.WatchedServices, db)
-						require.Contains(t, snap.TerminatingGateway.WatchedServices, billing)
-
-						require.Len(t, snap.TerminatingGateway.WatchedIntentions, 2)
-						require.Contains(t, snap.TerminatingGateway.WatchedIntentions, db)
-						require.Contains(t, snap.TerminatingGateway.WatchedIntentions, billing)
-
-						require.Len(t, snap.TerminatingGateway.WatchedLeaves, 2)
-						require.Contains(t, snap.TerminatingGateway.WatchedLeaves, db)
-						require.Contains(t, snap.TerminatingGateway.WatchedLeaves, billing)
-
-						require.Len(t, snap.TerminatingGateway.WatchedResolvers, 2)
-						require.Contains(t, snap.TerminatingGateway.WatchedResolvers, db)
-						require.Contains(t, snap.TerminatingGateway.WatchedResolvers, billing)
-
-						require.Len(t, snap.TerminatingGateway.GatewayServices, 2)
-						require.Contains(t, snap.TerminatingGateway.GatewayServices, db)
-						require.Contains(t, snap.TerminatingGateway.GatewayServices, billing)
-					},
-				},
-				verificationStage{
-					requiredWatches: map[string]verifyWatchRequest{
-						"external-service:db": genVerifyServiceWatch("db", "", "dc1", false),
-					},
-					events: []cache.UpdateEvent{
-						cache.UpdateEvent{
-							CorrelationID: "external-service:db",
-							Result: &structs.IndexedCheckServiceNodes{
-								Nodes: structs.CheckServiceNodes{
-									{
-										Node: &structs.Node{
-											Node:    "node1",
-											Address: "127.0.0.1",
-										},
-										Service: &structs.NodeService{
-											ID:      "db",
-											Service: "db",
-										},
-									},
-								},
-							},
-							Err: nil,
-						},
-					},
-					verifySnapshot: func(t testing.TB, snap *ConfigSnapshot) {
-						require.Len(t, snap.TerminatingGateway.ServiceGroups, 1)
-						require.Equal(t, snap.TerminatingGateway.ServiceGroups[structs.NewServiceID("db", nil)],
-							structs.CheckServiceNodes{
-								{
-									Node: &structs.Node{
-										Node:    "node1",
-										Address: "127.0.0.1",
-									},
-									Service: &structs.NodeService{
-										ID:      "db",
-										Service: "db",
-									},
-								},
-							},
-						)
-					},
-				},
-				verificationStage{
-					requiredWatches: map[string]verifyWatchRequest{
-						"service-leaf:db": genVerifyLeafWatch("db", "dc1"),
-					},
-					events: []cache.UpdateEvent{
-						cache.UpdateEvent{
-							CorrelationID: "service-leaf:db",
-							Result:        issuedCert,
-							Err:           nil,
-						},
-					},
-					verifySnapshot: func(t testing.TB, snap *ConfigSnapshot) {
-						require.Equal(t, snap.TerminatingGateway.ServiceLeaves[structs.NewServiceID("db", nil)], issuedCert)
-					},
-				},
-				verificationStage{
-					requiredWatches: map[string]verifyWatchRequest{
-						"service-resolver:db": genVerifyResolverWatch("db", "dc1", structs.ServiceResolver),
-					},
-					events: []cache.UpdateEvent{
-						cache.UpdateEvent{
-							CorrelationID: "service-resolver:db",
-							Result: &structs.IndexedConfigEntries{
-								Kind: structs.ServiceResolver,
-								Entries: []structs.ConfigEntry{
-									&structs.ServiceResolverConfigEntry{
-										Name: "db",
-										Kind: structs.ServiceResolver,
-										Redirect: &structs.ServiceResolverRedirect{
-											Service:    "db",
-											Datacenter: "dc2",
-										},
-									},
-								},
-							},
-							Err: nil,
-						},
-					},
-					verifySnapshot: func(t testing.TB, snap *ConfigSnapshot) {
-						want := &structs.ServiceResolverConfigEntry{
-							Kind: structs.ServiceResolver,
-							Name: "db",
-							Redirect: &structs.ServiceResolverRedirect{
-								Service:    "db",
-								Datacenter: "dc2",
-							},
-						}
-						require.Equal(t, want, snap.TerminatingGateway.ServiceResolvers[structs.NewServiceID("db", nil)])
-					},
-				},
-				verificationStage{
-					events: []cache.UpdateEvent{
-						cache.UpdateEvent{
-							CorrelationID: gatewayServicesWatchID,
-							Result: &structs.IndexedGatewayServices{
-								Services: structs.GatewayServices{
-									{
-										Service: structs.NewServiceID("billing", nil),
-										Gateway: structs.NewServiceID("terminating-gateway", nil),
-									},
-								},
-							},
-							Err: nil,
-						},
-					},
-					verifySnapshot: func(t testing.TB, snap *ConfigSnapshot) {
-						billing := structs.NewServiceID("billing", nil)
-
-						require.True(t, snap.Valid(), "gateway with service list is valid")
-
-						// All the watches should have been cancelled for db
-						require.Len(t, snap.TerminatingGateway.WatchedServices, 1)
-						require.Contains(t, snap.TerminatingGateway.WatchedServices, billing)
-
-						require.Len(t, snap.TerminatingGateway.WatchedIntentions, 1)
-						require.Contains(t, snap.TerminatingGateway.WatchedIntentions, billing)
-
-						require.Len(t, snap.TerminatingGateway.WatchedLeaves, 1)
-						require.Contains(t, snap.TerminatingGateway.WatchedLeaves, billing)
-
-						require.Len(t, snap.TerminatingGateway.WatchedResolvers, 1)
-						require.Contains(t, snap.TerminatingGateway.WatchedResolvers, billing)
-
-						require.Len(t, snap.TerminatingGateway.GatewayServices, 1)
-						require.Contains(t, snap.TerminatingGateway.GatewayServices, billing)
-
-						// There was no update event for billing's leaf/endpoints, so length is 0
-						require.Len(t, snap.TerminatingGateway.ServiceGroups, 0)
-						require.Len(t, snap.TerminatingGateway.ServiceLeaves, 0)
-						require.Len(t, snap.TerminatingGateway.ServiceResolvers, 0)
-					},
-				},
-			},
-		},
 		"connect-proxy":                    newConnectProxyCase(structs.MeshGatewayModeDefault),
 		"connect-proxy-mesh-gateway-local": newConnectProxyCase(structs.MeshGatewayModeLocal),
 	}
